Current spawner
The obstacleSpawner has a level up function so that if you want to change the levels, you can reset all the "balls" and set a new speed, speed increase per passing is +0.1. contains an arraylist of Obstacles with an x,y,radius,velocity that you can use for hit detection.